requirements.txt
openai
streamlit
PyPDF2
python-docx
pandas
langchain-openai
azure-search-documents
azure-core
pymupdf
openpyxl
fitz
st-pages
